\name{lgtdl}
\alias{lgtdl}
\title{Longitudinal Class }
\description{
  The \code{lgtdl} class is a simple class of R/S objects that make it
  simpler to manipulate longitudinal data.
}

\references{ Gentleman Robert (2001). Modeling with Objects.
  \emph{Proceedings of the 2nd International Workshop on Distributed
  Statistical Computing.
  \url{http://r-project.org/nosvn/conferences/DSC-2001/Proceedings/Gentleman.pdf}
}}

\author{Robert Gentleman}

\keyword{methods}
